Spike showing cube with TriangleMesh
We use the QuickHull3D library as a replacement for qhull (which svs_viewer
used). Figuring out the right data arrays to pass to TriangleMesh was a pain
in the neck, but it worked in the end! The next step will be to auto-convert
the vertices from SGEL using this code.