Add axes using FXyz's PolyLine3D
Create a new configurable builder for displaying axes. This required fiddling
around with lighting and materials for the rest of the scene because:

1. PolyLine3D (correctly, in my opinion) adds dedicated lighting for its lines,
2. JavaFX only adds default lighting when it detects that the scene has no
lights, and
3. because the lights in the PolyLine3D's are scoped to just those lines, the
rest of the scene was left completely black.

I couldn't figure out what the default lighting in JavaFX is, and Copilot tells
me it's not accessible and even that it might not be an accessible lighting
implementation at all. So I looked at svs_viewer's OpenGL settings and tried to
imitate them, but couldn't come up with the same thing. I find this low-contrast
and difficult to see.

I did an experiment and found that we could add visible lines on top of the
fill, so I'll come back to that.