label axes
I tried to add text into the 3d scene, but figuring that out was waaay too much
work. Changing the text to the right size affects its location! No thanks.
Overlaid labels are fine, thanks.

Extract label creation and relocation logic to a new Labels class. Put the label
style into the shared CSS.

Move the Axes to the GeometryManager, which needs to manage its label updating.
Encapsulate the axes in a dedicated class to hold its labels and root group.
It's unusual to put a class inside its builder instead of the other way around;
that should probably be fixed, but it can wait.